# docs/swagger.yaml
openapi: 3.0.0
info:
  title: FEDS Music API
  description: Swagger API documentation for user, musician, and admin endpoints authentication endpoint.
  version: 1.0.0
servers:
  - url: https://web-music-proj.onrender.com
paths:
...
  /admin/add_admin:
    post:
      summary: Register a new admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
              required: [username, password, email]
      responses:
        '200':
          description: Admin added successfully
  /admin/authenticate_admin:
    post:
      summary: Authenticate admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        '200':
          description: Admin authenticated successfully